mod: admin page structure
