+++ src/components/appRoutes.tsx
@@ -8,6 +8,10 @@ import NotFound from '../pages/notFound';
 import Subscription from '../pages/subscription';
 import Register from '../pages/register';
 import Delivery from '../pages/admin/delivery';
+import Profile from '../pages/profile';
+import AdminIndex from '../pages/admin/adminIndex';
+import AdminUsers from '../pages/admin/users';
+import AdminUser from '../pages/admin/user';
 
 const AppRoutes: React.FC = () => {
   return (
@@ -16,9 +20,15 @@ const AppRoutes: React.FC = () => {
       <Route path="/board" element={<Board />} />
       <Route path="/board/:id" element={<BoardDetail />} />
       <Route path="/login" element={<Login />} />
+      <Route path="/profile" element={<Profile />} />
       <Route path="/register" element={<Register />} />
-      <Route path="/admin/delivery" element={<Delivery />} />
       <Route path="/subscription" element={<Subscription />} />
+
+      {/* Admin Routes */}
+      <Route path="/admin" element={<AdminIndex />} />
+      <Route path="/admin/delivery" element={<Delivery />} />
+      <Route path="/admin/users" element={<AdminUsers />} />
+      <Route path="/admin/users/:id" element={<AdminUser />} />
       <Route path="*" element={<NotFound />} /> {/* 404 Not Found */}
     </Routes>
   );
